backend/requirements/base.txt
django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
pillow==12.0.0
